# Add performance tests to Pipelines, optimize enumerator performance a bit