--- a/scripts/build.ps1
+++ b/scripts/build.ps1
@@ -50,6 +50,12 @@ foreach ($file in [System.IO.Directory]::EnumerateFiles("$PSScriptRoot\..\tests"
         Write-Warning "Skipping tests in $file. These tests use the xunit.performance package which needs to be updated for dotnet SDK rc4."
         continue;
     }
+    if ($file -match "System.IO.Pipelines.Performance.Tests")
+    {
+        Write-Warning "Building Benchmark.NET project $file. Benchmark would not be ran as part of the build."
+        Invoke-Expression "$dotnetExePath build $file -c $Configuration"
+        continue;
+    }
     Write-Host "Building and running tests for project $file..."
     Invoke-Expression "$dotnetExePath test $file -c $Configuration -- -notrait category=performance -notrait category=outerloop"