--- a/tests/System.IO.Pipelines.Performance.Tests/configs/CoreConfig.cs
+++ b/tests/System.IO.Pipelines.Performance.Tests/configs/CoreConfig.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.IO.Pipelines.Performance.Tests.columns;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Engines;
+using BenchmarkDotNet.Jobs;
+using BenchmarkDotNet.Validators;
+
+namespace System.IO.Pipelines.Performance.Tests.configs
+{
+    public class CoreConfig : ManualConfig
+    {
+        public CoreConfig()
+        {
+            Add(JitOptimizationsValidator.FailOnError);
+            Add(new RpsColumn());
+
+            Add(Job.Default
+                .With(BenchmarkDotNet.Environments.Runtime.Core)
+                .WithRemoveOutliers(false)
+                .With(new GcMode() { Server = true })
+                .With(RunStrategy.Throughput)
+                .WithLaunchCount(3)
+                .WithWarmupCount(5)
+                .WithTargetCount(10));
+        }
+    }
+}