… src/System.IO.Pipelines/Properties/AssemblyInfo.cs