```
Before:

            Method |      Mean |    StdErr |    StdDev |    Median |        RPS | Allocated |
------------------ |---------- |---------- |---------- |---------- |----------- |---------- |
  MemoryEnumerator | 2.7179 us | 0.0419 us | 0.2296 us | 2.6041 us | 367,925.16 |       0 B |
 SegmentEnumerator | 1.6205 us | 0.0207 us | 0.1135 us | 1.6058 us | 617,082.52 |       0 B |

After:
            Method |      Mean |    StdErr |    StdDev |        RPS | Allocated |
------------------ |---------- |---------- |---------- |----------- |---------- |
  MemoryEnumerator | 2.4062 us | 0.0234 us | 0.1279 us | 415,591.36 |       0 B |
 SegmentEnumerator | 1.4740 us | 0.0178 us | 0.0974 us | 678,441.85 |       0 B |
```

@davidfowl @benaadams @KrzysztofCwalina