@benaadams 2017-02-23
`SegmentEnumerator` doesn't implement `IEnumerable` so ditch `Reset`?